feat: refactor code analysis state management and introduce repository layer
