+++ app/code_analysis/repositories/__init__.py
@@ -1,0 +1,3 @@
+"""
+Repository module for code analysis data access.
+"""
