+++ app/code_analysis/services/code_analysis.py
@@ -7,11 +7,9 @@ from logging import getLogger
 
 from bson import ObjectId
 
-from app.code_analysis.agents.code_analysis import (
-    create_code_analysis_agent,
-    get_analysis_state,
-)
+from app.code_analysis.agents.code_analysis import create_code_analysis_agent
 from app.code_analysis.models.code_analysis import CodeAnalysisState
+from app.code_analysis.repositories.code_analysis import get_analysis_state
 
 logger = getLogger(__name__)
